requirements.txt
asgiref==3.4.1
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.4.1
Django==3.2.9
django-rest-framework==0.1.0
djangorestframework==3.12.4
fonttools==4.27.1
html5lib==1.1
Pillow==8.4.0
pycparser==2.20
pydyf==0.1.2
pyphen==0.11.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.0
weasyprint==53.3
webencodings==0.5.1
zopfli==0.1.9
